RFCT(Move the controls for Rebar and Windows Ribbon into the main namespace, so I can make the constructors for 'test' versions non-public so they are not accidentally used in real code)
