+++ src/test/java/mmarquee/automation/controls/AutomationRibbonBarTest2.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package mmarquee.automation.controls.ribbon;
+package mmarquee.automation.controls;
 
 import com.sun.jna.Memory;
 import com.sun.jna.Native;
@@ -21,9 +21,8 @@ import com.sun.jna.Pointer;
 import com.sun.jna.ptr.PointerByReference;
 import mmarquee.automation.AutomationElement;
 import mmarquee.automation.ElementNotFoundException;
-import mmarquee.automation.UIAutomation;
+import mmarquee.automation.controls.AutomationRibbonBar;
 import mmarquee.automation.pattern.ItemContainer;
-import mmarquee.automation.uiautomation.IUIAutomation;
 import mmarquee.automation.uiautomation.IUIAutomationElement3;
 import org.junit.*;
 import org.mockito.Mockito;
@@ -32,7 +31,7 @@ import org.mockito.MockitoAnnotations;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.when;
@@ -40,9 +39,9 @@ import static org.mockito.Mockito.when;
 /**
  * Created by Mark Humphreys on 25/09/2017
  *
- * Tests for RibbonWorkPane controls
+ * Tests for AutomationRibbonBar.
  */
-public class AutomationRibbonWorkPaneTest2 {
+public class AutomationRibbonBarTest2 {
 
     @Before
     public void setup() {
@@ -63,27 +62,28 @@ public class AutomationRibbonWorkPaneTest2 {
     }
 
     @Test(expected = ElementNotFoundException.class)
-    public void testGetNUIPane_Throws_Exception_When_NUIPane_Not_Found() throws Exception {
+    public void testGetRibbonCommandBar_Throws_Exception_When_Not_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
         List<AutomationElement> collection = new ArrayList<>();
 
-        when(element.getClassName()).thenReturn(AutomationRibbonWorkPane.CLASS_NAME);
+        when(element.getClassName()).thenReturn(AutomationRibbonBar.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
         ItemContainer container = Mockito.mock(ItemContainer.class);
 
-        AutomationRibbonWorkPane workPane = new AutomationRibbonWorkPane(element, container);
+        AutomationRibbonBar bar = new AutomationRibbonBar(element, container);
 
-        workPane.getNUIPane(0);
+        bar.getRibbonCommandBar();
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
 
     @Test
-    public void testGetNUIPane_When_NUIPane_Is_Found() throws Exception {
+    public void testGetRibbonCommandBar_When_Element_Is_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
+        when(element.getClassName()).thenReturn(AutomationRibbonBar.CLASS_NAME);
         List<AutomationElement> collection = new ArrayList<>();
 
         IUIAutomationElement3 elem = Mockito.mock(IUIAutomationElement3.class);
@@ -93,23 +93,23 @@ public class AutomationRibbonWorkPaneTest2 {
                     Object[] args = invocation.getArguments();
                     PointerByReference reference = (PointerByReference) args[0];
 
-                    String value = "NUIPane";
+                    String value = "UIRibbonCommandBar";
                     Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
                     pointer.setWideString(0, value);
 
                     reference.setValue(pointer);
 
                     return 0;
-                });
+                }
+        );
 
         collection.add(new AutomationElement(elem));
 
-        when(element.getClassName()).thenReturn(AutomationRibbonWorkPane.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
-        AutomationRibbonWorkPane workPane = new AutomationRibbonWorkPane(element);
+        AutomationRibbonBar bar = new AutomationRibbonBar(element);
 
-        workPane.getNUIPane(0);
+        bar.getRibbonCommandBar();
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
