+++ src/test/java/mmarquee/automation/controls/AutomationRibbonCommandBarTest2.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package mmarquee.automation.controls.ribbon;
+package mmarquee.automation.controls;
 
 import com.sun.jna.Memory;
 import com.sun.jna.Native;
@@ -21,9 +21,9 @@ import com.sun.jna.Pointer;
 import com.sun.jna.ptr.PointerByReference;
 import mmarquee.automation.AutomationElement;
 import mmarquee.automation.ElementNotFoundException;
-import mmarquee.automation.UIAutomation;
+import mmarquee.automation.controls.AutomationRibbonCommandBar;
+import mmarquee.automation.controls.AutomationRibbonWorkPane;
 import mmarquee.automation.pattern.ItemContainer;
-import mmarquee.automation.uiautomation.IUIAutomation;
 import mmarquee.automation.uiautomation.IUIAutomationElement3;
 import org.junit.*;
 import org.mockito.Mockito;
@@ -40,13 +40,17 @@ import static org.mockito.Mockito.when;
 /**
  * Created by Mark Humphreys on 25/09/2017
  *
- * Tests for the NUIPane control.
+ * Tests for RibbonCommandBar
  */
-public class AutomationNUIPaneTest2 {
+public class AutomationRibbonCommandBarTest2 {
 
     @Before
     public void setup() {
         MockitoAnnotations.initMocks(this);
+    }
+
+    static {
+        ClassLoader.getSystemClassLoader().setDefaultAssertionStatus(true);
     }
 
     @BeforeClass
@@ -58,61 +62,55 @@ public class AutomationNUIPaneTest2 {
         return System.getProperty("os.name").toLowerCase().contains("windows");
     }
 
-    static {
-        ClassLoader.getSystemClassLoader().setDefaultAssertionStatus(true);
-    }
-
-    @Test(expected = ElementNotFoundException.class)
-    public void testGetNetUIHWND_Throws_Exception_When_NetUIHWND_Not_Found() throws Exception {
+    @Test (expected = ElementNotFoundException.class)
+    public void testGetRibbonCommandBar_Throws_Exception_When_No_WorkPane_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
-        when(element.getClassName()).thenReturn(AutomationNUIPane.CLASS_NAME);
         List<AutomationElement> collection = new ArrayList<>();
 
+        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
         ItemContainer container = Mockito.mock(ItemContainer.class);
 
-        AutomationNUIPane pane = new AutomationNUIPane(element, container);
+        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element, container);
 
-        pane.getNetUIHWND(0);
+        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
 
     @Test
-    public void testGetNetUIHWND_When_NetUIHWND_Is_Found() throws Exception {
+    public void testGetRibbonCommandBar_When_WorkPane_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
-        when(element.getClassName()).thenReturn(AutomationNUIPane.CLASS_NAME);
         List<AutomationElement> collection = new ArrayList<>();
 
         IUIAutomationElement3 elem = Mockito.mock(IUIAutomationElement3.class);
 
         Mockito.when(elem.getCurrentClassName(any())).thenAnswer(
-            invocation -> {
-                Object[] args = invocation.getArguments();
-                PointerByReference reference = (PointerByReference) args[0];
+                invocation -> {
+                    Object[] args = invocation.getArguments();
+                    PointerByReference reference = (PointerByReference) args[0];
 
-                String value = "NetUIHWND";
-                Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
-                pointer.setWideString(0, value);
+                    String value = "UIRibbonWorkPane";
+                    Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
+                    pointer.setWideString(0, value);
 
-                reference.setValue(pointer);
+                    reference.setValue(pointer);
 
-                return 0;
-            }
+                    return 0;
+                }
         );
 
         collection.add(new AutomationElement(elem));
 
+        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
-        ItemContainer container = Mockito.mock(ItemContainer.class);
+        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element);
 
-        AutomationNUIPane pane = new AutomationNUIPane(element, container);
-
-        pane.getNetUIHWND(0);
+        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
