+++ src/test/java/mmarquee/automation/controls/AutomationRibbonWorkPaneTest2.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package mmarquee.automation.controls.ribbon;
+package mmarquee.automation.controls;
 
 import com.sun.jna.Memory;
 import com.sun.jna.Native;
@@ -21,9 +21,8 @@ import com.sun.jna.Pointer;
 import com.sun.jna.ptr.PointerByReference;
 import mmarquee.automation.AutomationElement;
 import mmarquee.automation.ElementNotFoundException;
-import mmarquee.automation.UIAutomation;
+import mmarquee.automation.controls.AutomationRibbonWorkPane;
 import mmarquee.automation.pattern.ItemContainer;
-import mmarquee.automation.uiautomation.IUIAutomation;
 import mmarquee.automation.uiautomation.IUIAutomationElement3;
 import org.junit.*;
 import org.mockito.Mockito;
@@ -40,17 +39,13 @@ import static org.mockito.Mockito.when;
 /**
  * Created by Mark Humphreys on 25/09/2017
  *
- * Tests for RibbonCommandBar
+ * Tests for RibbonWorkPane controls
  */
-public class AutomationRibbonCommandBarTest2 {
+public class AutomationRibbonWorkPaneTest2 {
 
     @Before
     public void setup() {
         MockitoAnnotations.initMocks(this);
-    }
-
-    static {
-        ClassLoader.getSystemClassLoader().setDefaultAssertionStatus(true);
     }
 
     @BeforeClass
@@ -62,26 +57,30 @@ public class AutomationRibbonCommandBarTest2 {
         return System.getProperty("os.name").toLowerCase().contains("windows");
     }
 
-    @Test (expected = ElementNotFoundException.class)
-    public void testGetRibbonCommandBar_Throws_Exception_When_No_WorkPane_Found() throws Exception {
+    static {
+        ClassLoader.getSystemClassLoader().setDefaultAssertionStatus(true);
+    }
+
+    @Test(expected = ElementNotFoundException.class)
+    public void testGetNUIPane_Throws_Exception_When_NUIPane_Not_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
         List<AutomationElement> collection = new ArrayList<>();
 
-        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
+        when(element.getClassName()).thenReturn(AutomationRibbonWorkPane.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
         ItemContainer container = Mockito.mock(ItemContainer.class);
 
-        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element, container);
+        AutomationRibbonWorkPane workPane = new AutomationRibbonWorkPane(element, container);
 
-        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
+        workPane.getNUIPane(0);
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
 
     @Test
-    public void testGetRibbonCommandBar_When_WorkPane_Found() throws Exception {
+    public void testGetNUIPane_When_NUIPane_Is_Found() throws Exception {
         AutomationElement element = Mockito.mock(AutomationElement.class);
 
         List<AutomationElement> collection = new ArrayList<>();
@@ -93,24 +92,23 @@ public class AutomationRibbonCommandBarTest2 {
                     Object[] args = invocation.getArguments();
                     PointerByReference reference = (PointerByReference) args[0];
 
-                    String value = "UIRibbonWorkPane";
+                    String value = "NUIPane";
                     Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
                     pointer.setWideString(0, value);
 
                     reference.setValue(pointer);
 
                     return 0;
-                }
-        );
+                });
 
         collection.add(new AutomationElement(elem));
 
-        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
+        when(element.getClassName()).thenReturn(AutomationRibbonWorkPane.CLASS_NAME);
         when(element.findAll(any(), any())).thenReturn(collection);
 
-        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element);
+        AutomationRibbonWorkPane workPane = new AutomationRibbonWorkPane(element);
 
-        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
+        workPane.getNUIPane(0);
 
         Mockito.verify(element, atLeastOnce()).findAll(any(), any());
     }
