+++ src/test/java/mmarquee/automation/controls/AutomationWindowTest2.java
@@ -20,7 +20,6 @@ import com.sun.jna.platform.win32.Win32Exception;
 import com.sun.jna.platform.win32.WinDef;
 import com.sun.jna.ptr.IntByReference;
 import mmarquee.automation.*;
-import mmarquee.automation.controls.rebar.AutomationReBar;
 import mmarquee.automation.pattern.ItemContainer;
 import mmarquee.automation.pattern.PatternNotFoundException;
 import mmarquee.automation.pattern.Window;
